+++ x86/subops/x86_83.c
@@ -5,7 +5,7 @@ void ASM_83ADD(RM *rm, s32 disp, u8 val) {
     Reg res = { 0 };
     
     if (rm->isPtr) {
-        s64 fdisp = ASM_getDisp(rm, disp);
+        u32 fdisp = ASM_getDisp(rm, disp);
 
         switch (rm->otype) {
             case R_Bit16: { STACK16(temp, fdisp); prev.x = *temp; *temp += val; break; }
@@ -50,7 +50,7 @@ void ASM_83OR(RM *rm, s32 disp, u8 val) {
     Reg res = { 0 };
     
     if (rm->isPtr) {
-        s64 fdisp = ASM_getDisp(rm, disp);
+        u32 fdisp = ASM_getDisp(rm, disp);
         
         switch (rm->otype) {
             case R_Bit16: { STACK16(temp, fdisp); prev.x = *temp; *temp |= val; break; }
@@ -103,7 +103,7 @@ void ASM_83AND(RM *rm, s32 disp, u8 val) {
     s8 sval = (s8)val;
     
     if (rm->isPtr) {
-        s64 fdisp = ASM_getDisp(rm, disp);
+        u32 fdisp = ASM_getDisp(rm, disp);
         
         switch (rm->otype) {
             case R_Bit16: { STACK16(temp, fdisp); prev.x = *temp; *temp &= sval; break; }
@@ -154,7 +154,7 @@ void ASM_83SUB(RM *rm, s32 disp, u8 val) {
     Reg res = { 0 };
     
     if (rm->isPtr) {
-        s64 fdisp = ASM_getDisp(rm, disp);
+        u32 fdisp = ASM_getDisp(rm, disp);
 
         switch (rm->otype) {
             case R_Bit16: { STACK16(temp, fdisp); prev.x = *temp; *temp -= val; break; }
@@ -201,7 +201,7 @@ void ASM_83CMP(RM *rm, s32 disp, u8 val) {
     s8 sval = val;
     
     if (rm->isPtr) {
-        s64 fdisp = ASM_getDisp(rm, disp);
+        u32 fdisp = ASM_getDisp(rm, disp);
 
         switch (rm->otype) {
             case R_Bit16: { STACK16(temp, fdisp); prev.x = *temp; break; }
