+++ x86/subops/x86_C1.c
@@ -5,7 +5,7 @@ void ASM_C1SHL(RM *rm, s32 disp, u8 val) {
     Reg res = { 0 };
     
     if (rm->isPtr) {
-        s64 fdisp = ASM_getDisp(rm, disp);
+        u32 fdisp = ASM_getDisp(rm, disp);
 
         switch (rm->otype) {
             case R_Bit16: { STACK16(temp, fdisp); prev.x = *temp; *temp <<= val; break; }
@@ -52,7 +52,7 @@ void ASM_C1SHR(RM *rm, s32 disp, u8 val) {
     Reg res = { 0 };
     
     if (rm->isPtr) {
-        s64 fdisp = ASM_getDisp(rm, disp);
+        u32 fdisp = ASM_getDisp(rm, disp);
 
         switch (rm->otype) {
             case R_Bit16: { STACK16(temp, fdisp); prev.x = *temp; *temp >>= val; break; }
