+++ x86/subops/x86_F6.c
@@ -8,7 +8,7 @@ void ASM_F6TEST(RM *rm, s32 disp, u8 val) {
     Reg res = { 0 };
     
     if (rm->isPtr) {
-        s64 fdisp = ASM_getDisp(rm, disp);
+        u32 fdisp = ASM_getDisp(rm, disp);
 
         STACK8(temp, fdisp);
         prev.l = *temp;
